uwu yes uwu

+++ commands/admin/purge.js
@@ -4,26 +4,32 @@ module.exports = {
   name: 'purge',
   description: 'deletes the amount if specified messages',
   execute (client, message, args) {
-    const messageCount = args[0]
-    const messageCountInt = parseInt(messageCount)
-    const messageCountFinal = messageCountInt + 1
+    const messageCountFinal = parseInt(args[0]) + 1
 
     if (!message.member.permissions.has('MANAGE_CHANNELS')) {
       const eEmbed = new MessageEmbed()
-        .setColor('#ff1100')
+        .setColor('#FF0000')
         .setTitle('Delete Unsuccessfull')
-        .setDescription(`**${message.author.tag}** you dont have permission to use this command`)
+        .setDescription(
+          `${message.author} you dont have permission to use this command`)
         .setTimestamp()
-
       message.channel.send({ embeds: [eEmbed] })
     } else {
-      message.channel.bulkDelete(messageCountFinal)
-
-      if (messageCountFinal > 1) {
+      if (args[0] <= 0) {
         const eEmbed = new MessageEmbed()
-          .setColor('#00ff04')
+          .setColor('#FF0000')
+          .setTitle('Delete Unsuccessfull')
+          .setDescription(
+          `${message.author} I can not delete ${args[0]} messages`
+          )
+          .setTimestamp()
+        message.channel.send({ embeds: [eEmbed] })
+      } else {
+        message.channel.bulkDelete(messageCountFinal)
+        const eEmbed = new MessageEmbed()
+          .setColor('#00FF00')
           .setTitle('Messages Deleted')
-          .setDescription(`Deleted ${messageCount} messages`)
+          .setDescription(`Deleted ${args[0]} messages`)
         message.channel.send({ embeds: [eEmbed] })
       }
     }
